+++ CHANGELOG.md
@@ -16,10 +16,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Added interface `DateTime`.
 - Added class `LocalDateTime`. (A date-time without a time-zone in the
   ISO-8601 calendar system, such as 2007-12-03T10:15:30).
-- Added `Location` and `LocationId`.
-- Added `Locations` - repository of locations.
+- Added `TimeZone` and `TimeZoneId`.
+- Added `TimeZones` - repository of time zones.
 - Added `ZonedDateTime`.
-- Added `Zone`.
+- Added `Offset`.
 - Added interface `Format\DateTimeFormatter`.
 - Added interface `Format\DateTimeParser`.
 - Added interface `Format\DateFormatter`.